Add cast to int to failureError.code so this test works on both the iOS and Mac OS X frameworks.

+++ Specs/Network/RKRequestSpec.m
@@ -97,7 +97,7 @@
     [[[loaderMock expect] andForwardToRealObject] request:request didFailLoadWithError:OCMOCK_ANY];
     [request sendAsynchronously];
     [loaderMock waitForResponse];
-    assertThatInt(loader.failureError.code, equalToInt(RKRequestConnectionTimeoutError));
+    assertThatInt((int)loader.failureError.code, equalToInt(RKRequestConnectionTimeoutError));
     [request release];
 }
